\name{AMPSequences}
\alias{AMPSequences}
\docType{data}
\title{
Antimicrobial peptide sequences
}
\description{
  Sequences corresponding to the \code{\link{AMPSequences.response}}
  dataset, which contains the antimicrobial activity of the sequences.
}
\usage{data(AMPSequences)}
\format{
  The data is an object of class \code{\linkS4class{Sequences}}.
}

\source{ Nathan Ostberg, Yiannis Kaznessis, "Protegrin structure-activity
  relationships: using homology modelsp of synthetic sequences to
  determine structural characteristics important for activity", Peptides,
  Volume 26, Issue 2, February 2005, Pages 197-206, ISSN 0196-9781,
  10.1016/j.peptides.2004.09.020.  }

\keyword{datasets}
