\name{AMPSequences.response}
\alias{AMPSequences.response}
\docType{data}
\title{
Antimicrobial peptide sequence responses
}
\description{
  A data frame containing the antimicrobial activity for each peptide in
  the \code{\link{AMPSequences}} dataset.
}
\usage{data(AMPSequences.response)}
\format{
  \describe{
    \item{\code{Sequence}}{The sequence name}
    \item{\code{E-coli}}{The activity on organism E. coli}
    \item{\code{N-gonorrhoeae-F.62.}}{The activity on organism N. gonorrhoeae}
    \item{\code{N-gonorrhoeae-FA.19.}}{The activity on organism N. gonorrhoeae}
    \item{\code{L-monocytogenes}}{The activity on organism L. monocytogenes}
    \item{\code{C-albicans}}{The activity on organism C. albicans}
    \item{\code{P-aeruginosa}}{The activity on organism P. aeruginosa}
    \item{\code{Cytotoxicity}}{Cytotoxicity of peptide}
    \item{\code{Hemolysis}}{Negative effect on red blood cell}
  }
}


  \source{ Nathan Ostberg, Yiannis Kaznessis, "Protegrin structure-activity
  relationships: using homology modelsp of synthetic sequences to
  determine structural characteristics important for activity", Peptides,
  Volume 26, Issue 2, February 2005, Pages 197-206, ISSN 0196-9781,
  10.1016/j.peptides.2004.09.020.  }

\keyword{datasets}
