\name{default.MetricParams}
\alias{default.MetricParams}
\docType{data}
\title{
Default parameters for calculating sequence-sequence distances
}
\description{
Consists of the BLOSUM62 substitution matrix and a gap open penalty of
10 and gap extension penalty of 0.2. These are used when comparing sequences.
}
\format{
  An object of class \code{\linkS4class{MetricParams}}
}
\references{
Henikoff, S.; Henikoff, J.G. (1992). "Amino Acid Substitution Matrices
from Protein Blocks". PNAS 89 (22): 10915-10919.
}

\keyword{datasets}
