\name{SHP2Sequences}
\alias{SHP2Sequences}
\docType{data}
\title{
  SHP-2_N-SH2 peptide sequences
}
\description{
  Peptide sequences where the peptides are known to be de-phosphorylated
  on a phosphorylated tyrosine residue in the third position of the
  sequences.
}
\usage{data(SHP2Sequences)}
\format{
    The data is an object of class \code{\linkS4class{Sequences}}.
}

\source{

Decoding Protein-Protein Interactions through Combinatorial
Chemistry: Sequence Specificity of SHP-1, SHP-2, and SHIP SH2 Domains
Michael C. Sweeney Anne-Sophie Wavreille, Junguk Park, Jonathan
P. Butchar, Susheela Tridandapani, and, and Dehua Pei,
Biochemistry 2005 44 (45), 14932-14947 }

\keyword{datasets}
