\name{TULASequences}
\alias{TULASequences}
\docType{data}
\title{
  TULA-2_Post peptide sequences
}
\description{
  Peptide sequences where the peptides are known to be de-phosphorylated
  on a phosphorylated tyrosine residue just before the first
  position. These are the sequences that come after that residue.
}
\usage{data(SHP2Sequences)}
\format{
    The data is an object of class \code{\linkS4class{Sequences}}.
}

\source{
Chen, X.; Ren, L.; Kim, S.; Carpino, N.; Daniel, J. L.; Kunapuli, S. P.;
Tsygankov, A. Y.; Pei, D. The Journal of biological chemistry 2010, 285,
31268-31276.
}

\keyword{datasets}
